refactor(cockpit): extract review packet comment renderer
